backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
Pillow
openpyxl==3.1.2
pandas==2.1.3
python-decouple==3.8
requests==2.31.0
celery==5.3.4
redis==5.0.1
django-filter==23.3
drf-yasg==1.21.7
